Use NUnit 3, VS2015 minor tweaks for filepaths
